// src/typings/api/assets/assets.d.ts
export type WithdrawAddressListResp = {
  /** 备注 */
  remark?: string;
  id: number;
  /** 地址 */
  address: string
}
// ...
export type AssetsWithdrawSubmitReq = {
  /** 币种 ID */
  coinId: string;
  /** 提币地址，链上转账必传 */
  address?: string;
  /** 币种符号 */
  symbol: string;
  /** 提币数量 */
  amount: number;
  /** 充值地址标签，如果需要 memo，此处为必填 */
  memo?: string;
  /** 1.链上转账，3，内部转账 */
  type: number;
  /** 用户唯一的 uid，内部转账必传 */
  uid?: string;
  /** 接口唯一参数 */
  uuid: string;
}
// ...
export interface ISafeVerifySendWithdrawalDataProps {
  address: string // 提币地址
  quantity: string | number // 提币数量
  currencyCode: string // 提币币种
  memo: string // memo 地址
}